Initial commit moving project code into fresh repo
We are just about to port over to a BOOTBOOT loader system and x64, which will
require a bit of an overhaul. Since gullfoss seems to be headed in a fun
direction, I figured it was worth splitting out into its own repository.

The next few steps will mostly be making lateral progress until x64 support is
in place properly.