Adds the barest frame of a wallpaper script
Using a wlroots analogue, wallpaper will need to interact with a `layer-shell`
surface. We need to build out the whole shell / surface / message system in
Haskell types. Ultimately the idea in this VM test is that the wayland graphics
system will get statically linked into the VM, but will dynamically load a set
of independent "startup" programs, including things like the wallpaper. The
dynamic load still runs this code in threads with shared memory, so the wayland
system can establish shared channels, etc with these graphical clients and
implement some wayland-like protocol for clients to draw into the graphics
system. The system in turn speaks to the simple openGL virtual "display" and
ensures all the buffers and drawn out in the right order.