Starts to formalize process interface, adds some signal handling
The emulator interface to processes is begining to be defined by exposed
kernel core/module functions that are extern declared in a kernel.rs lib.
The graphics "driver" also has a similar interface library in the userspace.
It's still unclear how we can expose these without needing to compile unsafe
code.